chromefill: add auto transpiling capabilities
Adds auto transpiling using Babel standalone for nullish coalescing
and optional chaining, since these can't be polyfilled.
This functionality is very slow and CPU and memory intensive. It's not pretty,
and only intended as a last resort. As of now, this support is experimental
and whitelisted on a per-site basis in the extension itself.